refactor: reorganize and update DTO imports and remove unused components
