+++ src/modules/repos/http/dto/get-user-repositories-params.ts
@@ -1,0 +1,6 @@
+export interface GetUserRepositoriesParams {
+  username: string;
+  authorizationToken?: string;
+  page: number;
+  perPage?: number;
+}
