+++ src/modules/repos/http/dto/user-repositories-data.ts
@@ -1,0 +1,8 @@
+export interface UserRepositoriesData {
+  id: number;
+  name: string;
+  description: string | null;
+  owner: {
+    login: string;
+  };
+}
