+++ src/modules/shared/services/github-service.ts
@@ -1,7 +1,7 @@
 import { getFollowingUsers } from '@/modules/following/http/get-following-users';
 import { createRepository } from '@/modules/repos/http/create-repository';
 import { getUserRepositories } from '@/modules/repos/http/get-user-repositories';
-import { getUser } from '../http/get-user';
+import { getUser } from '../../user/http/get-user';
 import type { LuizaHubService } from '../types/luiza-hub';
 
 export class GithubService implements LuizaHubService {
