+++ src/modules/shared/services/mock-luizahub-service.ts
@@ -2,7 +2,7 @@ import { followingMock } from '@/mocks/following-mock';
 import { repositoriesMock } from '@/mocks/repositories-mock';
 import type { GetFollowingUsersParams } from '@/modules/following/http/dto/get-following-users-params';
 import type { CreateRepositoryParams } from '@/modules/repos/http/dto/create-repository-params';
-import type { GetUserRepositoriesParams } from '@/modules/repos/http/get-user-repositories';
+import type { GetUserRepositoriesParams } from '@/modules/repos/http/dto/get-user-repositories-params';
 import type { LuizaHubService } from '@/modules/shared/types/luiza-hub';
 
 export class MockLuizaHubService implements LuizaHubService {
