+++ src/modules/shared/types/luiza-hub.ts
@@ -1,7 +1,7 @@
 import type { FollowingUserData } from '@/modules/following/http/dto/following-user-data';
 import type { CreatedRepositoryData } from '@/modules/repos/http/dto/created-repository-data';
-import type { UserRepositoriesData } from '@/modules/repos/http/get-user-repositories';
-import type { UserResponseData } from '../http/dto/get-user-response-data';
+import type { UserRepositoriesData } from '@/modules/repos/http/dto/user-repositories-data';
+import type { UserResponseData } from '../../user/http/dto/get-user-response-data';
 import type { PaginatedResult } from './paginated-result';
 
 export interface LuizaHubService {
